Extract the route for employee
