+++ app.rb
@@ -6,6 +6,7 @@ require 'twilio-ruby'
 require_relative 'helpers/data_mapper_setup'
 require_relative 'models/employee'
 require_relative 'lib/searcher'
+require_relative 'routes/employee'
 
 ENV['RACK_ENV'] ||= 'development'
 
@@ -17,50 +18,7 @@ class EmployeeDirectoryApp < Sinatra::Application
   config_file 'config/app.yml.erb'
   Helpers::DataMapperSetup.setup(settings.database_url)
 
-  get '/employee' do
-    employees = Searcher.search(params[:Body])
-
-    content_type 'application/xml'
-
-    case employees.size
-    when 0 then employee_not_found_message
-    when 1 then details_message(employees.first)
-    else listing_message(employees)
-    end
-  end
-
-  private
-
-  def employee_not_found_message
-    Twilio::TwiML::Response.new do |response|
-      response.Message 'not found'
-    end.to_xml
-  end
-
-  def listing_message(employees)
-    Twilio::TwiML::Response.new do |response|
-      response.Message employees_label_list(employees).join(' ')
-    end.to_xml
-  end
-
-  def employees_label_list(employees)
-    employees.map do |employee|
-      "#{employee.id}-#{employee.name}"
-    end
-  end
-
-  def details_message(employee)
-    Twilio::TwiML::Response.new do |response|
-      employee_info = "#{employee.id}-#{employee.name}"\
-        " #{employee.email}"\
-        " #{employee.phone_number}"
-
-      response.Message do |message|
-        message.Body employee_info
-        message.Media employee.image_url
-      end
-    end.to_xml
-  end
+  register Routes::Employee
 
   run! if app_file == $PROGRAM_NAME
 end
